+++ components/VoteResultStatus.tsx
@@ -3,7 +3,7 @@ import { BanIcon } from '@heroicons/react/solid'
 import useProposal from '@hooks/useProposal'
 import useProposalVotes from '@hooks/useProposalVotes'
 import { ProposalState } from '@solana/spl-governance'
-import { useVetoingPop } from './VotePanel'
+import { useVetoingPop } from './VotePanel/VotePanel'
 
 const VetoResult = () => {
   const vetoingPop = useVetoingPop()
