+++ pages/dao/[symbol]/proposal/[pk]/index.tsx
@@ -5,7 +5,7 @@ import useProposal from 'hooks/useProposal'
 import ProposalStateBadge from '@components/ProposalStateBadge'
 import { InstructionPanel } from 'components/instructions/instructionPanel'
 import DiscussionPanel from 'components/chat/DiscussionPanel'
-import VotePanel from 'components/VotePanel'
+import VotePanel from '@components/VotePanel/VotePanel'
 import { ApprovalProgress, VetoProgress } from '@components/QuorumProgress'
 import useRealm from 'hooks/useRealm'
 import useProposalVotes from 'hooks/useProposalVotes'
